+++ addon/utils/should-add-token.js
@@ -1,36 +1,24 @@
-export default function shouldAddToken (url, portalId = '', portalHost) {
-  // if this is a hosted service that belongs to the user's organization
-  // if this is a service with stored credentials
-  const id = portalId.toLowerCase();
-  if (portalHost) {
-    return hostsMatch(portalHost, url);
-  } else {
-    return (
-      !!url.toLowerCase().match('arcgis.com/arcgis/rest/services') ||
-      !!url.toLowerCase().match(`arcgis.com/${id}/arcgis/rest`) ||
-      /utility\w+\.arcgis\.com\/sharing\/servers\/\w{32}\/rest\/services\/Hosted/i.test(url)
-    );
-  }
+export default function shouldAddToken (url, serverInfo, portalInfo) {
+  const serverDomain = stripToDomain(url);
+  const portalDomain = stripToDomain(portalInfo.portalHostname);
+  const owningDomain = stripToDomain(serverInfo.owningSystemUrl);
+  const authorizedCrossOriginDomains = portalInfo.authorizedCrossOriginDomains || [];
+  const isAuthorizedUrl = authorizedCrossOriginDomains.includes(serverDomain);
+  const isArcGisDomain = !!url.toLowerCase().match('.arcgis.com/');
+
+  if (serverDomain === portalDomain === owningDomain) return true;
+  else if (portalDomain === owningDomain && isAuthorizedUrl) return true;
+  else if (portalDomain === owningDomain && isArcGisDomain) return true;
+  return false;
 }
 
-export function hostsMatch (currentHost, requestedUrl) {
+function stripToDomain (url) {
   try {
-    const portalTopTwo = stripToTopDomain(currentHost);
-    const requestedTopTwo = stripToTopDomain(requestedUrl);
-    return portalTopTwo === requestedTopTwo;
+    return url
+    .replace(/^https?:\/\//, '')
+    .split(':')[0]
+    .split('/')[0];
   } catch (e) {
-    return false;
+    return url;
   }
-}
-
-function stripToTopDomain (url) {
-  // pop off arcgis.com from server.foo.bar.arcgis.com/foobarbaz
-  return url
-  .replace(/^https?:\/\//, '')
-  .split(':')[0]
-  .split('/')[0]
-  .split('')
-  .reverse()
-  .join('')
-  .match(/^[^.]+\.[^.]+/)[0];
 }