+++ tests/dummy/config/environment.js
@@ -36,12 +36,18 @@ module.exports = function (environment) {
   ENV.torii.providers['arcgis-oauth-bearer'].portalUrl = ENV.APP.portalBaseUrl;
 
   if (environment === 'development') {
-    const env = 'dev';
-    ENV.APP.domainServiceUrl = 'https://servicesdev.arcgis.com/LjjARY1mkhxulWPq/arcgis/rest/services/sitedomains/FeatureServer/0';
-    ENV.torii.providers['arcgis-oauth-bearer'].apiKey = 'ifq94vTWyyZclwNz';
-    // if you need to test against QA API/data:
-    // const env = 'qa';
-    // ENV.APP.domainServiceUrl = 'https://servicesqa.arcgis.com/97KLIFOSt5CxbiRI/arcgis/rest/services/sitedomains/FeatureServer/0';
+    // const env = 'dev';
+    const env = 'qa';
+
+    switch (env) {
+      case 'qa':
+        ENV.APP.domainServiceUrl = 'https://servicesqa.arcgis.com/97KLIFOSt5CxbiRI/arcgis/rest/services/sitedomains/FeatureServer/0';
+        ENV.torii.providers['arcgis-oauth-bearer'].apiKey = 'Ll5erY6niWgZS1eV';
+        break;
+      default:
+        ENV.APP.domainServiceUrl = 'https://servicesdev.arcgis.com/LjjARY1mkhxulWPq/arcgis/rest/services/sitedomains/FeatureServer/0';
+        ENV.torii.providers['arcgis-oauth-bearer'].apiKey = 'ifq94vTWyyZclwNz';
+    }
 
     ENV.torii.providers['arcgis-oauth-bearer'].portalUrl = `https://${env}ext.arcgis.com`;
   }
@@ -58,7 +64,6 @@ module.exports = function (environment) {
     ENV.torii.providers['arcgis-oauth-bearer'].apiKey = 'Ll5erY6niWgZS1eV';
     ENV.torii.providers['arcgis-oauth-bearer'].portalUrl = 'https://qaext.arcgis.com';
     ENV.APP.domainServiceUrl = 'https://servicesqa.arcgis.com/97KLIFOSt5CxbiRI/arcgis/rest/services/sitedomains/FeatureServer/0';
-
   }
 
   if (environment === 'test') {